requirements.txt
easyocr==1.7.1
Flask==3.0.3
joblib==1.4.2
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
numpy==1.26.4
open-clip-torch==2.24.0
openai-clip==1.0.1
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
pandas==2.2.2
pillow==10.3.0
requests==2.31.0
scikit-learn==1.2.2
torch==2.3.0